requirements.txt
Django>=4.1.7,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.26.1,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
django-cors-headers>=3.14.0,<3.15